Added checks for data type of parameters for resources of tasks

+++ src/extra/model.py
@@ -134,14 +134,14 @@ class AlibabaModelDist(SyntheticModelDist):
             results_data_size = self.results_scaling * rnd.uniform(*self.results_range)
             foreknowledge_task = Task(
                 f'Foreknowledge Task {task_id}',
-                required_storage=self.storage_scaling * task_row['mem-max'],
-                required_computation=self.computational_scaling * task_row['cpu-avg'] * task_row['time-taken'],
+                required_storage=int(self.storage_scaling * task_row['mem-max']),
+                required_computation=int(self.computational_scaling * task_row['cpu-avg'] * task_row['time-taken']),
                 required_results_data=ceil(results_data_size * task_row['mem-max']),
                 deadline=task_row['time-taken'], servers=servers)
             requested_task = Task(
                 f'Requested Task {task_id}',
-                required_storage=self.storage_scaling * task_row['request-mem'],
-                required_computation=self.computational_scaling * task_row['request-cpu'] * task_row['time-taken'],
+                required_storage=int(self.storage_scaling * task_row['request-mem']),
+                required_computation=int(self.computational_scaling * task_row['request-cpu'] * task_row['time-taken']),
                 required_results_data=ceil(results_data_size * task_row['request-mem']),
                 deadline=task_row['time-taken'], value=foreknowledge_task.value)
